# internal/docs/swagger.yaml
basePath: /
definitions:
...
  internal_modules_user_controller.User:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    required:
    - email
    - firstName
    - lastName
    - password
    - phone
    - username
    type: object
  internal_modules_user_controller.UserArray:
    properties:
      user:
        items:
          $ref: '#/definitions/internal_modules_user_controller.User'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    Зоомагазин, который работает с тремя основными сущностями: пользователь, животное и магазин.
    Здесь доступны CRUD-операции над всеми сущностями, мягкая работа с пользователем (проставление дат действий), авторизация и выход из системы с помощью чёрного списка, загрузка изображения и сохранение локально на сервере. Это самые интересные возможности, попробуйте сами
  title: Petstore
  version: "1.0"
paths:
...
  /user/:
    post:
      consumes:
      - application/json
      description: Создание нового пользователя или восстановление удалённого с различными
        полями. Отсчёт будет идти от username (не должен повторяться)
      parameters:
      - description: Заполни все поля для регистрации
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_modules_user_controller.User'
      produces:
      - text/plain
      responses:
        "200":
          description: Успешная регистрация или не ошибочное сообщение
          schema:
            type: string
        "400":
          description: Неверной формат структуры
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Регистрация нового пользователя
      tags:
      - user
  /user/{username}:
    delete:
      consumes:
      - application/json
      description: Мягкое удаление пользователя (проставление даты)
      parameters:
      - description: Укажи username
        in: path
        name: username
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Успешное удаление или не ошибочное сообщение
          schema:
            type: string
        "400":
          description: Неверной формат
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Удаление пользователя
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Получить полную информацию о пользователе
      parameters:
      - description: Укажи username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение информации или не ошибочное сообщение
          schema:
            $ref: '#/definitions/internal_modules_user_controller.User'
        "400":
          description: Неверной формат
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Получить информацию о пользователе
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Нельзя обновить на то имя пользователя, которое уже есть в базе
      parameters:
      - description: Укажи username
        in: path
        name: username
        required: true
        type: string
      - description: Заполни все поля для обновления информации
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_modules_user_controller.User'
      produces:
      - text/plain
      responses:
        "200":
          description: Успешное обновление информации или не ошибочное сообщение
          schema:
            type: string
        "400":
          description: Неверной формат
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Обновить информацию о пользователе
      tags:
      - user
  /user/createWithArray:
    post:
      consumes:
      - application/json
      description: Создание новой группы пользователей с различными полями. Отсчёт
        будет идти от username (не должен повторяться). Для восстановления аккаунта
        воспользуйтесь одиночной регистрацией
      parameters:
      - description: Заполни все поля для регистрации
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_modules_user_controller.UserArray'
      produces:
      - text/plain
      responses:
        "200":
          description: Успешная регистрация или не ошибочное сообщение
          schema:
            type: string
        "400":
          description: Неверной формат структуры
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Регистрация группы пользователей
      tags:
      - user
  /user/login:
    get:
      consumes:
      - application/json
      description: Создание токена jwt и отправка в заголовке для дальнейшей авторизации
      parameters:
      - description: Укажи username
        in: query
        name: username
        required: true
        type: string
      - description: Укажи пароль
        in: query
        name: password
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Успешный вход в систему или не ошибочное сообщение
          schema:
            type: string
        "400":
          description: Неверной формат
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Вход в систему
      tags:
      - user
  /user/logout:
    get:
      consumes:
      - application/json
      description: Аннулирование jwt токена с использованием black_list
      parameters:
      - description: Укажи username
        in: query
        name: username
        required: true
        type: string
      - description: Укажи пароль
        in: query
        name: password
        required: true
        type: string
      - description: Bearer токен доступа для имитации отправки браузером. Обязательно
          укажите Bearer перед вставкой токена
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Успешный выход из системы или не ошибочное сообщение
          schema:
            type: string
        "400":
          description: Неверной формат
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Выход из системы
      tags:
      - user
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"