# internal/modules/user/controller/users.go
package controller

type User struct {
	UserName  string `json:"username" validate:"required"`
	FirstName string `json:"firstName" validate:"required"`
	LastName  string `json:"lastName" validate:"required"`
	Email     string `json:"email" validate:"required"`
	Password  string `json:"password" validate:"required"`
	Phone     string `json:"phone" validate:"required"`
}

type UserArray struct {
	UserArray []User `json:"user"`
}